Remove Nightly dependencies of Criterion use
For reasons not yet explained, if I use the test::black_box to wrap
top level Criterion benchmarks I get differing (reduced) performance
vs the built-in black_box of Criterion.  But this forces Nightly use
in `cargo test` which I do not want. Falling back to Criterion
black_box now but noting that it deviates.