Add complement function for Intervals
Initial implementation attempt wsa to return an iterator over Intervals,
but that seemed to prove sticky to implement without overhead (Rust needs
to know the return value size - if I am to return by value).

For now I've settled on a 2-tuple in which the first value is always present
(as every Interval complement yields at least one interval) and the second
Interval returned in the tuple is Optional (not always present, depends
upon the input Interval variant).